+++ windows/StealthMessage/Views/HubView.xaml.cs
@@ -1,0 +1,8 @@
+using Microsoft.UI.Xaml.Controls;
+
+namespace StealthMessage.Views;
+
+public sealed partial class HubView : UserControl
+{
+    public HubView() => InitializeComponent();
+}